refactor chat app
refactor chat into multiple smaller components
